# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalOut": {
        "additionalProperties": false,
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseOut"
          },
          "case_created": {
            "description": "False when an active case already covered this pair.",
            "title": "Case Created",
            "type": "boolean"
          },
          "match": {
            "$ref": "#/components/schemas/MatchOut"
          }
        },
        "required": [
          "match",
          "case",
          "case_created"
        ],
        "title": "ApprovalOut",
        "type": "object"
      },
...
      "BulkIn": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "description": "Approval is deliberately absent: it opens a case, one record at a time.",
            "enum": [
              "reject",
              "escalate"
            ],
            "title": "Action",
            "type": "string"
          },
          "comment": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Comment",
            "type": "string"
          },
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids",
          "action",
          "comment"
        ],
        "title": "BulkIn",
        "type": "object"
      },
      "BulkItemOut": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBodyOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          }
        },
        "required": [
          "id",
          "ok"
        ],
        "title": "BulkItemOut",
        "type": "object"
      },
      "BulkOut": {
        "additionalProperties": false,
        "properties": {
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BulkItemOut"
            },
            "title": "Results",
            "type": "array"
          },
          "succeeded": {
            "title": "Succeeded",
            "type": "integer"
          }
        },
        "required": [
          "succeeded",
          "failed",
          "results"
        ],
        "title": "BulkOut",
        "type": "object"
      },
...
      "CommentIn": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Comment",
            "type": "string"
          },
          "provider_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          }
        },
        "required": [
          "comment"
        ],
        "title": "CommentIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchDetailOut": {
        "additionalProperties": false,
        "description": "Everything the Investigation screen renders.",
        "properties": {
          "adjudication": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdjudicationOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "The LLM's answer, re-validated from the stored response, when an LLM decided."
          },
          "approved_provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Provider Id"
          },
          "band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BandOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "The thresholds of the scoring config that decided this result."
          },
          "calibrated_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Confidence"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseOut"
            },
            "title": "Cases",
            "type": "array"
          },
          "chosen_provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Provider Id"
          },
          "conflicting_cases": {
            "description": "Live cases opened on an earlier result that this result disagrees with (Q5).",
            "items": {
              "$ref": "#/components/schemas/CaseOut"
            },
            "title": "Conflicting Cases",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "explanation": {
            "additionalProperties": true,
            "description": "Reason code, notes, and the adjudicator's reasoning and evidence when an LLM decided.",
            "title": "Explanation",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "llm": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider, model and prompt version of the adjudication call.",
            "title": "Llm"
          },
          "posterior": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posterior"
          },
          "raw_match_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Match Weight"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By Email"
          },
          "reviewer_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Comment"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sanction_record": {
            "$ref": "#/components/schemas/SanctionRecordDetailOut"
          },
          "sanction_record_id": {
            "format": "uuid",
            "title": "Sanction Record Id",
            "type": "string"
          },
          "superseded_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By"
          }
        },
        "required": [
          "id",
          "run_id",
          "sanction_record_id",
          "decision",
          "chosen_provider_id",
          "approved_provider_id",
          "posterior",
          "calibrated_confidence",
          "raw_match_weight",
          "route",
          "review_status",
          "reviewed_by",
          "reviewed_at",
          "reviewer_comment",
          "superseded_by",
          "created_at",
          "explanation",
          "candidates",
          "sanction_record"
        ],
        "title": "MatchDetailOut",
        "type": "object"
      },
...
      "Page_MatchListItemOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchListItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MatchListItemOut]",
        "type": "object"
      },
...
      "ReviewIn": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "duration_months": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Case length on approval; default 3.",
            "title": "Duration Months"
          },
          "provider_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required to approve an ambiguous match; must be one of its candidates.",
            "title": "Provider Id"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "title": "ReviewIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Provider sanctions and exclusions reconciliation.\n\nMatches sanction records against a provider master file with a Fellegi-Sunter\nengine, EM-learned weights and isotonic calibration, and routes the grey band to\nan LLM adjudicator. Every run is pinned to a data snapshot and can be replayed.",
    "title": "Concordance",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/matches": {
      "get": {
        "description": "Current results, filtered and sorted, each beside the record it decides.",
        "operationId": "list_matches_matches_get",
        "parameters": [
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size.",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "MATCH",
                    "AMBIGUOUS",
                    "NO_MATCH"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "review_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PENDING",
                    "APPROVED",
                    "REJECTED",
                    "ESCALATED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Status"
            }
          },
          {
            "in": "query",
            "name": "min_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Confidence"
            }
          },
          {
            "in": "query",
            "name": "max_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Confidence"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "sanction_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sanction Type"
            }
          },
          {
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "individual",
                    "organization"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Record Type"
            }
          },
          {
            "description": "Only results involved in a conflict on a live case.",
            "in": "query",
            "name": "conflict",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only results involved in a conflict on a live case.",
              "title": "Conflict"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "description": "Include results a later run replaced (Q5).",
            "in": "query",
            "name": "include_superseded",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include results a later run replaced (Q5).",
              "title": "Include Superseded",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "confidence",
              "enum": [
                "confidence",
                "date"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MatchListItemOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Matches",
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/bulk": {
      "post": {
        "description": "Reject or escalate up to 100 results with one comment.\n\nEach item is decided and audited exactly as the single endpoint would, in\nits own savepoint: a refused item is reported beside its id and the rest\nstill commit. The response is 200 even when some items fail - read\n`failed`. Approval is not offered in bulk.",
        "operationId": "bulk_matches_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bulk",
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/{match_id}": {
      "get": {
        "description": "Everything behind one decision: candidates, per-field evidence, the AI's reasoning.\n\nRecorded in the audit log as a view.",
        "operationId": "get_match_matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Match",
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/{match_id}/approve": {
      "post": {
        "description": "Confirm the match and open its case. Admin only.",
        "operationId": "approve_matches__match_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Already approved or rejected, or superseded."
          },
          "422": {
            "description": "An ambiguous match needs `provider_id`, from its candidates."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/{match_id}/escalate": {
      "post": {
        "description": "Hand the decision to an admin, saying why.",
        "operationId": "escalate_matches__match_id__escalate_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Escalate",
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/{match_id}/reject": {
      "post": {
        "description": "The proposed provider is not the sanctioned party. A comment is required.",
        "operationId": "reject_matches__match_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject",
        "tags": [
          "matches"
        ]
      }
    },
...
  }
}